requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.6.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Vector DB
pinecone-client>=3.0.0

# Embeddings
sentence-transformers>=2.6.1
torch>=2.2.0

# LLMs
langchain-core>=0.2.0
langchain-groq>=0.1.5

# Utilities
tqdm>=4.66.0
